initial commit of the saas jira source plugin
