--- data-prepper-connector-sdk/src/main/java/org/opensearch/dataprepper/connector/sdk/base/BaseConnectorSourceConfig.java
@@ -1,6 +1,0 @@
-package org.opensearch.dataprepper.connector.sdk.base;
-
-/**
- * Marker interface to all the SAAS connectors configuration
- */
-public interface BaseConnectorSourceConfig { }
