--- data-prepper-connector-sdk/src/main/java/org/opensearch/dataprepper/connector/sdk/coordination/LeaderProgressState.java
@@ -1,19 +1,0 @@
-package org.opensearch.dataprepper.connector.sdk.coordination;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Setter
-@Getter
-public class LeaderProgressState {
-
-    @JsonProperty("initialized")
-    private boolean initialized = false;
-
-    @JsonProperty("streamArns")
-    private List<String> streamArns;
-
-}
