+++ data-prepper-main/src/main/resources/META-INF/data-prepper.plugins.properties
@@ -3,4 +3,4 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-org.opensearch.dataprepper.plugin.packages=org.opensearch.dataprepper.plugins,org.opensearch.dataprepper.connector,com.amazon.dataprepper.plugins
+org.opensearch.dataprepper.plugin.packages=org.opensearch.dataprepper.plugins,com.amazon.dataprepper.plugins
