+++ data-prepper-plugins/build.gradle
@@ -13,5 +13,6 @@ subprojects {
 }
 
 dependencies {
-    subprojects.forEach { api project(':data-prepper-plugins:' + it.name) }
+    subprojects.findAll { api project(it.path) }
 }
+
