+++ data-prepper-plugins/saas-source-plugins/saas-crawler/build.gradle
@@ -1,0 +1,31 @@
+plugins {
+    id 'java-library'
+}
+
+group = 'org.opensearch.dataprepper.plugins.source.saas.crawler'
+
+tasks.withType(Javadoc).configureEach {
+    enabled = false
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+    implementation project(path: ':data-prepper-api')
+
+    implementation 'com.fasterxml.jackson.core:jackson-core'
+    implementation 'com.fasterxml.jackson.core:jackson-databind'
+    implementation 'org.projectlombok:lombok:1.18.30'
+
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
+
+    testImplementation platform('org.junit:junit-bom:5.10.0')
+    testImplementation 'org.junit.jupiter:junit-jupiter'
+}
+
+test {
+    useJUnitPlatform()
+}
