+++ data-prepper-plugins/saas-source-plugins/saas-crawler/src/main/java/org/opensearch/dataprepper/plugins/source/saas/crawler/base/BaseConnectorSourceConfig.java
@@ -1,0 +1,6 @@
+package org.opensearch.dataprepper.plugins.source.saas.crawler.base;
+
+/**
+ * Marker interface to all the SAAS connectors configuration
+ */
+public interface BaseConnectorSourceConfig { }
