+++ data-prepper-plugins/saas-source-plugins/saas-crawler/src/main/java/org/opensearch/dataprepper/plugins/source/saas/crawler/coordination/LeaderProgressState.java
@@ -1,0 +1,19 @@
+package org.opensearch.dataprepper.plugins.source.saas.crawler.coordination;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Setter
+@Getter
+public class LeaderProgressState {
+
+    @JsonProperty("initialized")
+    private boolean initialized = false;
+
+    @JsonProperty("streamArns")
+    private List<String> streamArns;
+
+}
